services/inbound_dock_lambda/requirements.txt
psycopg2-binary==2.9.9
boto3==1.35.0
feedparser==6.0.11
python-dateutil==2.9.0
